batch: Extract hunk replacement translation
Live-hunk ownership translation now has a separate hunk_line_ranges scanner
and a shared LineEntry helper for presence baseline references.

The translator still owns file-derived replacement-run pre-consumption,
which mixes range scanning, absence content capture, consumed IDs, and
ownership assembly in the same function.

This commit moves replacement-run translation into
hunk_replacement_translation and has the translator merge the returned
range-backed ownership pieces with the remaining hunk walk.

Subsequent commits will add direct replacement-phase coverage and an
import-boundary guard.
